fixed particle localization issue for quadratic and gaussian methods which turned out to be because of integer intensity values for particles and gaussian optimization was not solving because of garbage initial values

+++ include/typedefs.h
@@ -99,7 +99,7 @@ struct particle2d {
     double x;
     double y;
     double z;
-    int I;
+    double I;
     int size;
 };
